+++ jessy-cabs/src/component/Info/RateManagement/PackageRateEntery/PackageRateEntery.jsx
@@ -68,6 +68,8 @@ const columns = [
 
 
 const PackageRateEntery = () => {
+  const user_id = localStorage.getItem('useridno');
+
   const [selectedCustomerData, setSelectedCustomerData] = useState({});
   const [selectedCustomerId, setSelectedCustomerId] = useState(null);
   const [rows, setRows] = useState([]);
@@ -78,9 +80,63 @@ const PackageRateEntery = () => {
   const [warning, setWarning] = useState(false);
   const [successMessage, setSuccessMessage] = useState({});
   const [errorMessage, setErrorMessage] = useState({});
-  const [warningMessage] = useState({});
+  const [warningMessage, setWarningMessage] = useState({});
   const [infoMessage] = useState({});
 
+
+  // for page permission
+
+  const [userPermissions, setUserPermissions] = useState({});
+
+  useEffect(() => {
+    const fetchPermissions = async () => {
+      try {
+        const currentPageName = 'Rate Type';
+        const response = await axios.get(`http://localhost:8081/user-permissions/${user_id}/${currentPageName}`);
+        setUserPermissions(response.data);
+        console.log('permission data', response.data);
+      } catch (error) {
+        console.error('Error fetching user permissions:', error);
+      }
+    };
+
+    fetchPermissions();
+  }, [user_id]);
+
+  const checkPagePermission = () => {
+    const currentPageName = 'Rate Type';
+    const permissions = userPermissions || {};
+
+    if (permissions.page_name === currentPageName) {
+      return {
+        read: permissions.read_permission === 1,
+        new: permissions.new_permission === 1,
+        modify: permissions.modify_permission === 1,
+        delete: permissions.delete_permission === 1,
+      };
+    }
+
+    return {
+      read: false,
+      new: false,
+      modify: false,
+      delete: false,
+    };
+  };
+
+  const permissions = checkPagePermission();
+
+  // Function to determine if a field should be read-only based on permissions
+  const isFieldReadOnly = (fieldName) => {
+    if (permissions.read) {
+      // If user has read permission, check for other specific permissions
+      if (fieldName === "delete" && !permissions.delete) {
+        return true;
+      }
+      return false;
+    }
+    return true;
+  };
 
 
   const hidePopup = () => {
@@ -214,21 +270,31 @@ const PackageRateEntery = () => {
     setSelectedCustomerData(customerData);
     setSelectedCustomerId(params.row.customerId);
   }, []);
+
   const handleAdd = async () => {
-    const duty = book.duty;
-    if (!duty) {
-      setError(true);
-      setErrorMessage("Check your Network Connection");
-      return;
-    }
-    try {
-      await axios.post('http://localhost:8081/ratemanagement', book);
-      handleCancel();
-      setSuccess(true);
-      setSuccessMessage("Successfully Added");
-    } catch {
-      setError(true);
-      setErrorMessage("Check your Network Connection");
+    const permissions = checkPagePermission();
+
+    if (permissions.read && permissions.new) {
+      const duty = book.duty;
+      if (!duty) {
+        setError(true);
+        setErrorMessage("Check your Network Connection");
+        return;
+      }
+      try {
+        await axios.post('http://localhost:8081/ratemanagement', book);
+        handleCancel();
+        setRows([]);
+        setSuccess(true);
+        setSuccessMessage("Successfully Added");
+      } catch {
+        setError(true);
+        setErrorMessage("Check your Network Connection");
+      }
+    } else {
+      // Display a warning or prevent the action
+      setWarning(true);
+      setWarningMessage("You do not have permission.");
     }
   };
 
@@ -236,32 +302,55 @@ const PackageRateEntery = () => {
     event.preventDefault();
     try {
       if (actionName === 'List') {
-        const response = await axios.get('http://localhost:8081/ratemanagement');
-        const data = response.data;
-        if (data.length > 0) {
-          setRows(data);
-          setSuccess(true);
-          setSuccessMessage("Successfully listed");
+        const permissions = checkPagePermission();
+
+        if (permissions.read && permissions.read) {
+          const response = await axios.get('http://localhost:8081/ratemanagement');
+          const data = response.data;
+          if (data.length > 0) {
+            setRows(data);
+            setSuccess(true);
+            setSuccessMessage("Successfully listed");
+          } else {
+            setRows([]);
+            setError(true);
+            setErrorMessage("No data found");
+          }
         } else {
-          setRows([]);
-          setError(true);
-          setErrorMessage("No data found");
+          setWarning(true);
+          setWarningMessage("You do not have permission.");
         }
       } else if (actionName === 'Cancel') {
         handleCancel();
       } else if (actionName === 'Delete') {
-        await axios.delete(`http://localhost:8081/ratemanagement/${selectedCustomerData.id}`);
-        setSelectedCustomerData(null);
-        setSuccess(true);
-        setSuccessMessage("Successfully Deleted");
-        handleCancel();
+        const permissions = checkPagePermission();
+
+        if (permissions.read && permissions.delete) {
+          await axios.delete(`http://localhost:8081/ratemanagement/${selectedCustomerData.id}`);
+          setSelectedCustomerData(null);
+          setSuccess(true);
+          setSuccessMessage("Successfully Deleted");
+          handleCancel();
+          setRows([]);
+        } else {
+          setWarning(true);
+          setWarningMessage("You do not have permission.");
+        }
       } else if (actionName === 'Edit') {
-        const selectedCustomer = rows.find((row) => row.id === selectedCustomerData.id);
-        const updatedCustomer = { ...selectedCustomer, ...selectedCustomerData };
-        await axios.put(`http://localhost:8081/ratemanagement/${selectedCustomerData.id}`, updatedCustomer);
-        setError(true);
-        setSuccessMessage("Successfully updated");
-        handleCancel();
+        const permissions = checkPagePermission();
+
+        if (permissions.read && permissions.modify) {
+          const selectedCustomer = rows.find((row) => row.id === selectedCustomerData.id);
+          const updatedCustomer = { ...selectedCustomer, ...selectedCustomerData };
+          await axios.put(`http://localhost:8081/ratemanagement/${selectedCustomerData.id}`, updatedCustomer);
+          setError(true);
+          setSuccessMessage("Successfully updated");
+          handleCancel();
+          setRows([]);
+        } else {
+          setWarning(true);
+          setWarningMessage("You do not have permission.");
+        }
       } else if (actionName === 'Add') {
         handleAdd();
       }
@@ -581,7 +670,7 @@ const PackageRateEntery = () => {
               />
             </div>
             <div className="input" style={{ width: "100px" }}>
-              <Button variant="contained" onClick={handleAdd}>Save All</Button>
+              <Button variant="contained" onClick={handleAdd} disabled={isFieldReadOnly("new")}>Save All</Button>
             </div>
           </div>
         </div>
